feat(cli): integrate max tracing requests in cli (#2914)

+++ crates/rpc/rpc/src/call_guard.rs
@@ -4,14 +4,14 @@ use tokio::sync::{AcquireError, OwnedSemaphorePermit, Semaphore};
 /// RPC Tracing call guard semaphore.
 ///
 /// This is used to restrict the number of concurrent RPC requests to tracing methods like
-/// `debug_traceTransaction` because they can consume a lot of memory.
+/// `debug_traceTransaction` because they can consume a lot of memory and CPU.
 #[derive(Clone, Debug)]
 pub struct TracingCallGuard(Arc<Semaphore>);
 
 impl TracingCallGuard {
     /// Create a new `TracingCallGuard` with the given maximum number of tracing calls in parallel.
-    pub fn new(max_tracing_requests: usize) -> Self {
-        Self(Arc::new(Semaphore::new(max_tracing_requests)))
+    pub fn new(max_tracing_requests: u32) -> Self {
+        Self(Arc::new(Semaphore::new(max_tracing_requests as usize)))
     }
 
     /// See also [Semaphore::acquire_owned]
